+++ packages/patrol/android/src/main/kotlin/pl/leancode/patrol/PatrolJUnitRunner.java
@@ -35,8 +35,6 @@ public class PatrolJUnitRunner extends AndroidJUnitRunner {
     public PatrolAppServiceClient patrolAppServiceClient;
     private Map<String, Boolean> dartTestCaseSkipMap = new HashMap<>();
 
-    private String spaceReplacement = "__";
-
     @Override
     protected boolean shouldWaitForActivitiesToComplete() {
         return false;
@@ -117,9 +115,8 @@ public class PatrolJUnitRunner extends AndroidJUnitRunner {
             List<DartGroupEntry> dartTestCases = ContractsExtensionsKt.listTestsFlat(dartTestGroup, "");
             List<String> dartTestCaseNamesList = new ArrayList<>();
             for (DartGroupEntry dartTestCase : dartTestCases) {
-                final String testName = sanitizeTestCaseName(dartTestCase.getName());
-                dartTestCaseSkipMap.put(testName, dartTestCase.getSkip());
-                dartTestCaseNamesList.add(testName);
+                dartTestCaseSkipMap.put(dartTestCase.getName(), dartTestCase.getSkip());
+                dartTestCaseNamesList.add(dartTestCase.getName());
             }
             Object[] dartTestCaseNames = dartTestCaseNamesList.toArray();
             Logger.INSTANCE.i(TAG + "Got Dart tests: " + Arrays.toString(dartTestCaseNames));
@@ -145,7 +142,7 @@ public class PatrolJUnitRunner extends AndroidJUnitRunner {
 
         try {
             Logger.INSTANCE.i(TAG + "Requested execution");
-            RunDartTestResponse response = patrolAppServiceClient.runDartTest(originalTestCaseName(name));
+            RunDartTestResponse response = patrolAppServiceClient.runDartTest(name);
             if (response.getResult() == Contracts.RunDartTestResponseResult.failure) {
                 throw new AssertionError("Dart test failed: " + name + "\n" + response.getDetails());
             }
@@ -155,20 +152,5 @@ public class PatrolJUnitRunner extends AndroidJUnitRunner {
             Logger.INSTANCE.e(TAG + e.getMessage(), e.getCause());
             throw new RuntimeException(e);
         }
-    }
-
-    /**
-     * We need to remove whitespaces from test case name in order to make in compatible with Orchestrator 1.5.0.
-     * New requirement can be observed (<a href="https://github.com/android/android-test/commit/8383d784e51dd67972f79f7738e19e7e99706d23">here</a>).
-     * */
-    private String sanitizeTestCaseName(String name) {
-        return name.replace(" ", spaceReplacement);
-    }
-
-    /**
-     * When calling test on dart side, we need to bring back original test case name.
-    * */
-    private String originalTestCaseName(String name) {
-        return name.replace(spaceReplacement, " ");
     }
 }