+++ packages/patrol_cli/CHANGELOG.md
@@ -1,4 +1,4 @@
-## 3.5.0
+## Unreleased
 
 - Add android product flavor to dart-define. (#2425)