+++ packages/patrol_cli/CHANGELOG.md
@@ -1,6 +1,7 @@
-## Unreleased
+## 3.4.1
 
 - Add android product flavor to dart-define. (#2425)
+- Detect and warn about Orchestrator 1.5.0. (#2437) 
 
 ## 3.4.0